Moves perform_changes() from object_list to ScimActions.
